requirements.txt
PySocks==1.6.8
